+++ src/nvim/viml/dumpers/dumpers.c
@@ -111,7 +111,7 @@ int dump_unumber(const uintmax_t unumber, Writer write, void *cookie)
   char result[MAXNUMBUFLEN];
   char *e = result;
   sdump_unumber(unumber, &e);
-  return write_string_len(result, e - result, write, cookie);
+  return write_string_len(result, (size_t) (e - result), write, cookie);
 }
 
 #define ABS(n) ((uintmax_t) (n >= 0 ? n : -n))
@@ -147,7 +147,7 @@ int dump_number(const intmax_t number, Writer write, void *cookie)
   char result[MAXNUMBUFLEN + 1];
   char *e = result;
   sdump_number((intmax_t) number, &e);
-  return write_string_len(result, e - result, write, cookie);
+  return write_string_len(result, (size_t) (e - result), write, cookie);
 }
 
 #undef ABS
